services/vanna/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-dotenv==1.0.1
vanna==0.7.9
psycopg[binary]==3.2.1
pandas==2.2.2
chromadb==0.5.5
pydantic==2.9.0
